generic encoder pipeline
